move piechart component
